When augmenting device, find hidden devices and flag them as hidden
